+++ Labs/5. Logging and Diagnostics.md
@@ -22,7 +22,7 @@
     }
     ```
     
-1. In Visual Studio, change the active launch host to the application itself (self-host) by navigating to the play/run button and changing the drop-down to the entry named after the application. 
+1. In Visual Studio, change the active launch host to the application itself (self-host) by navigating to the play/run button and changing the drop-down to the entry named after the application.
 
       ![image](Images/run-with-kestrel.png)
