+++ Labs/Code/Lab5A/src/Lab5A/Program.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 
-namespace Lab4
+namespace Lab5A
 {
     public class Program
     {
