+++ Labs/Code/Lab5B/Lab5B.sln
@@ -1,11 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26020.0
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.28917.182
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{14917E84-0349-44AA-9E50-EECB3DAB9976}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Lab4", "src\Lab4\Lab4.csproj", "{3B55DFDA-79ED-4938-903F-5ED05C76ACB7}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Lab5B", "src\Lab5B\Lab5B.csproj", "{3B55DFDA-79ED-4938-903F-5ED05C76ACB7}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -37,6 +37,6 @@ Global
 		{3B55DFDA-79ED-4938-903F-5ED05C76ACB7} = {14917E84-0349-44AA-9E50-EECB3DAB9976}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {BBA096C2-42FC-405B-BEC2-CB4972578542}
+		SolutionGuid = {9A62DE30-9367-4A8B-AF28-8747C3D645A0}
 	EndGlobalSection
 EndGlobal
