+++ Labs/Code/Lab5B/src/Lab5B/Program.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 
-namespace Lab4
+namespace Lab5B
 {
     public class Program
     {
