--- Labs/Code/Lab5C/src/Lab4/logfile.txt
@@ -1,16 +1,0 @@
-2019-05-12 15:22:59.347 +02:00 [INF] Application startup complete!
-2019-05-12 15:22:59.376 +02:00 [FTL] This is a critical message
-2019-05-12 15:22:59.399 +02:00 [WRN] This is a warning message
-2019-05-12 15:22:59.401 +02:00 [ERR] This is an error message
-2019-05-12 15:23:08.243 +02:00 [INF] Request starting HTTP/1.1 GET http://localhost:8081/  
-2019-05-12 15:23:10.440 +02:00 [ERR] An unhandled exception has occurred while executing the request.
-System.InvalidOperationException: Oops!
-   at Lab4.Startup.<>c.<Configure>b__6_1(HttpContext context) in C:\Sources\arnaudmaichac\aspnetcore-concepts-workshop\Labs\Code\Lab4C\src\Lab4\Startup.cs:line 63
-   at Microsoft.AspNetCore.Diagnostics.ExceptionHandlerMiddleware.Invoke(HttpContext context)
-2019-05-12 15:23:10.488 +02:00 [INF] Request finished in 2251.089ms 500 text/html
-2019-05-12 15:23:10.585 +02:00 [INF] Request starting HTTP/1.1 GET http://localhost:8081/favicon.ico  
-2019-05-12 15:23:11.492 +02:00 [ERR] An unhandled exception has occurred while executing the request.
-System.InvalidOperationException: Oops!
-   at Lab4.Startup.<>c.<Configure>b__6_1(HttpContext context) in C:\Sources\arnaudmaichac\aspnetcore-concepts-workshop\Labs\Code\Lab4C\src\Lab4\Startup.cs:line 63
-   at Microsoft.AspNetCore.Diagnostics.ExceptionHandlerMiddleware.Invoke(HttpContext context)
-2019-05-12 15:23:11.500 +02:00 [INF] Request finished in 912.8695ms 500 text/html
