+++ Labs/Code/Lab5C/src/Lab5C/Program.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 
-namespace Lab4
+namespace Lab5C
 {
     public class Program
     {
