+++ Labs/Code/Lab5C/src/Lab5C/Startup.cs
@@ -1,13 +1,17 @@
-﻿using Microsoft.AspNetCore.Builder;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-using Serilog;
+using Microsoft.Extensions.Configuration;
 using System.IO;
+using Serilog;
 
-namespace Lab4
+namespace Lab5C
 {
     public class Startup
     {
@@ -41,11 +45,22 @@ namespace Lab4
             {
                 app.UseDeveloperExceptionPage();
             }
-
-            app.Run(async (context) =>
+            else
             {
-                //await context.Response.WriteAsync($"Hello World! {env.EnvironmentName}");
-                await context.Response.WriteAsync($"{Configuration["message"]}");
+                app.UseExceptionHandler(subApp =>
+                {
+                    subApp.Run(async context =>
+                    {
+                        context.Response.ContentType = "text/html";
+                        await context.Response.WriteAsync("<strong> Application error. Please contact support. </strong>");
+                        await context.Response.WriteAsync(new string(' ', 512));  // Padding for IE
+                    });
+                });
+            }
+
+            app.Run((context) =>
+            {
+                throw new InvalidOperationException("Oops!");
             });
 
             startupLogger.LogInformation("Application startup complete!");
