+++ Labs/Code/Lab5D/Lab5D.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 15.0.26020.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{14917E84-0349-44AA-9E50-EECB3DAB9976}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Lab4", "src\Lab4\Lab4.csproj", "{3B55DFDA-79ED-4938-903F-5ED05C76ACB7}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Lab5D", "src\Lab5D\Lab5D.csproj", "{3B55DFDA-79ED-4938-903F-5ED05C76ACB7}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -37,6 +37,6 @@ Global
 		{3B55DFDA-79ED-4938-903F-5ED05C76ACB7} = {14917E84-0349-44AA-9E50-EECB3DAB9976}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {9A62DE30-9367-4A8B-AF28-8747C3D645A0}
+		SolutionGuid = {BB0BD81B-B05D-435E-B267-6FF9F74E13D2}
 	EndGlobalSection
 EndGlobal
