+++ Labs/Code/Lab5D/src/Lab5D/Program.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 
-namespace Lab4
+namespace Lab5D
 {
     public class Program
     {
