+++ Labs/Code/Lab5D/src/Lab5D/logfile.txt
@@ -1,0 +1,92 @@
+2019-05-12 15:28:52.917 +02:00 [INF] Application startup complete!
+2019-05-12 15:28:52.999 +02:00 [FTL] This is a critical message
+2019-05-12 15:28:53.004 +02:00 [WRN] This is a warning message
+2019-05-12 15:28:53.007 +02:00 [ERR] This is an error message
+2019-05-12 15:29:02.320 +02:00 [INF] Request starting HTTP/1.1 GET http://localhost:8081/  
+2019-05-12 15:29:02.356 +02:00 [INF] Request finished in 38.8702ms 404 text/html
+2019-05-12 15:29:02.501 +02:00 [INF] Request starting HTTP/1.1 GET http://localhost:8081/favicon.ico  
+2019-05-12 15:29:02.512 +02:00 [INF] Request finished in 6.5073ms 404 text/html
+2019-06-27 13:22:23.500 -04:00 [INF] Application startup complete!
+2019-06-27 13:22:23.528 -04:00 [FTL] This is a critical message
+2019-06-27 13:22:23.534 -04:00 [WRN] This is a warning message
+2019-06-27 13:22:23.540 -04:00 [ERR] This is an error message
+2019-06-27 13:22:24.083 -04:00 [INF] Request starting HTTP/1.1 GET http://localhost:55740/  
+2019-06-27 13:22:24.083 -04:00 [INF] Request starting HTTP/1.1 DEBUG http://localhost:55740/  0
+2019-06-27 13:22:24.133 -04:00 [INF] Request finished in 54.2099ms 404 
+2019-06-27 13:22:24.133 -04:00 [INF] Request finished in 54.0921ms 200 
+2019-06-27 13:27:46.128 -04:00 [INF] Application startup complete!
+2019-06-27 13:27:46.151 -04:00 [FTL] This is a critical message
+2019-06-27 13:27:46.156 -04:00 [WRN] This is a warning message
+2019-06-27 13:27:46.159 -04:00 [ERR] This is an error message
+2019-06-27 13:27:46.610 -04:00 [INF] Request starting HTTP/1.1 DEBUG http://localhost:55740/  0
+2019-06-27 13:27:46.610 -04:00 [INF] Request starting HTTP/1.1 GET http://localhost:55740/  
+2019-06-27 13:27:46.794 -04:00 [INF] Request finished in 186.6284ms 404 
+2019-06-27 13:27:46.794 -04:00 [INF] Request finished in 186.6197ms 200 
+2019-06-27 13:28:33.386 -04:00 [INF] Application startup complete!
+2019-06-27 13:28:33.413 -04:00 [FTL] This is a critical message
+2019-06-27 13:28:33.418 -04:00 [WRN] This is a warning message
+2019-06-27 13:28:33.426 -04:00 [ERR] This is an error message
+2019-06-27 13:28:33.805 -04:00 [INF] Request starting HTTP/1.1 GET http://localhost:55740/  
+2019-06-27 13:28:33.805 -04:00 [INF] Request starting HTTP/1.1 DEBUG http://localhost:55740/  0
+2019-06-27 13:28:33.858 -04:00 [INF] Request finished in 57.5073ms 404 
+2019-06-27 13:28:33.858 -04:00 [INF] Request finished in 57.5281ms 200 
+2019-06-27 13:28:57.202 -04:00 [INF] Application startup complete!
+2019-06-27 13:28:57.233 -04:00 [FTL] This is a critical message
+2019-06-27 13:28:57.241 -04:00 [WRN] This is a warning message
+2019-06-27 13:28:57.244 -04:00 [ERR] This is an error message
+2019-06-27 13:28:58.033 -04:00 [INF] Request starting HTTP/1.1 DEBUG http://localhost:55740/  0
+2019-06-27 13:28:58.033 -04:00 [INF] Request starting HTTP/1.1 GET http://localhost:55740/  
+2019-06-27 13:28:58.150 -04:00 [INF] Request finished in 93.4583ms 404 
+2019-06-27 13:28:58.150 -04:00 [INF] Request finished in 93.8655ms 200 
+2019-06-27 13:31:32.002 -04:00 [INF] Application startup complete!
+2019-06-27 13:31:32.027 -04:00 [FTL] This is a critical message
+2019-06-27 13:31:32.029 -04:00 [WRN] This is a warning message
+2019-06-27 13:31:32.031 -04:00 [ERR] This is an error message
+2019-06-27 13:31:32.598 -04:00 [INF] Request starting HTTP/1.1 DEBUG http://localhost:55740/  0
+2019-06-27 13:31:32.598 -04:00 [INF] Request starting HTTP/1.1 GET http://localhost:55740/  
+2019-06-27 13:31:32.665 -04:00 [INF] Request finished in 68.4719ms 200 
+2019-06-27 13:31:32.665 -04:00 [INF] Request finished in 68.4719ms 404 
+2019-06-27 13:34:32.846 -04:00 [INF] Application startup complete!
+2019-06-27 13:34:32.867 -04:00 [FTL] This is a critical message
+2019-06-27 13:34:32.867 -04:00 [WRN] This is a warning message
+2019-06-27 13:34:32.867 -04:00 [ERR] This is an error message
+2019-06-27 13:34:33.096 -04:00 [INF] Request starting HTTP/1.1 DEBUG http://localhost:55740/  0
+2019-06-27 13:34:33.096 -04:00 [INF] Request starting HTTP/1.1 DEBUG http://localhost:55740/  0
+2019-06-27 13:34:33.096 -04:00 [INF] Request starting HTTP/1.1 GET http://localhost:55740/  
+2019-06-27 13:34:33.118 -04:00 [INF] Request finished in 24.9985ms 404 
+2019-06-27 13:34:33.119 -04:00 [INF] Request finished in 25.0166ms 404 
+2019-06-27 13:34:33.118 -04:00 [INF] Request finished in 24.9904ms 404 
+2019-06-27 13:35:11.828 -04:00 [INF] Application startup complete!
+2019-06-27 13:35:11.854 -04:00 [FTL] This is a critical message
+2019-06-27 13:35:11.856 -04:00 [WRN] This is a warning message
+2019-06-27 13:35:11.858 -04:00 [ERR] This is an error message
+2019-06-27 13:35:12.234 -04:00 [INF] Request starting HTTP/1.1 GET http://localhost:55740/  
+2019-06-27 13:35:12.234 -04:00 [INF] Request starting HTTP/1.1 DEBUG http://localhost:55740/  0
+2019-06-27 13:35:12.282 -04:00 [INF] Request finished in 50.5549ms 404 
+2019-06-27 13:35:12.282 -04:00 [INF] Request finished in 50.5468ms 200 
+2019-06-27 13:35:17.337 -04:00 [INF] Request starting HTTP/1.1 GET http://localhost:55740/  
+2019-06-27 13:35:17.340 -04:00 [INF] Request finished in 2.6848ms 404 
+2019-06-27 13:35:48.894 -04:00 [INF] Application startup complete!
+2019-06-27 13:35:48.925 -04:00 [FTL] This is a critical message
+2019-06-27 13:35:48.927 -04:00 [WRN] This is a warning message
+2019-06-27 13:35:48.936 -04:00 [ERR] This is an error message
+2019-06-27 13:35:49.344 -04:00 [INF] Request starting HTTP/1.1 DEBUG http://localhost:55740/  0
+2019-06-27 13:35:49.344 -04:00 [INF] Request starting HTTP/1.1 GET http://localhost:55740/  
+2019-06-27 13:35:49.402 -04:00 [INF] Request finished in 61.4686ms 200 
+2019-06-27 13:35:49.402 -04:00 [INF] Request finished in 61.4697ms 404 
+2019-06-27 13:36:52.222 -04:00 [INF] Application startup complete!
+2019-06-27 13:36:52.238 -04:00 [FTL] This is a critical message
+2019-06-27 13:36:52.238 -04:00 [WRN] This is a warning message
+2019-06-27 13:36:52.238 -04:00 [ERR] This is an error message
+2019-06-27 13:36:52.583 -04:00 [INF] Request starting HTTP/1.1 GET http://localhost:55740/  
+2019-06-27 13:36:52.583 -04:00 [INF] Request starting HTTP/1.1 DEBUG http://localhost:55740/  0
+2019-06-27 13:36:52.603 -04:00 [INF] Request finished in 22.6349ms 404 
+2019-06-27 13:36:52.603 -04:00 [INF] Request finished in 22.6501ms 404 
+2019-06-27 13:37:53.996 -04:00 [INF] Application startup complete!
+2019-06-27 13:37:54.020 -04:00 [FTL] This is a critical message
+2019-06-27 13:37:54.022 -04:00 [WRN] This is a warning message
+2019-06-27 13:37:54.025 -04:00 [ERR] This is an error message
+2019-06-27 13:37:54.602 -04:00 [INF] Request starting HTTP/1.1 GET http://localhost:55740/  
+2019-06-27 13:37:54.603 -04:00 [INF] Request starting HTTP/1.1 DEBUG http://localhost:55740/  0
+2019-06-27 13:37:58.072 -04:00 [INF] Request finished in 107.4866ms 200 
+2019-06-27 13:37:58.085 -04:00 [INF] Request finished in 3516.8353ms 404 text/html
